An iterator over &HyperplaneMethod that produces [T; D].
A problem: the `Precomputed` variant will produce values with a
lifetime of 'a (hopefully 'static), but the `OnDemand` variant will
produce values of `b (which are on the function call stack?). To
avoid this, make a clone of the values. There's a TODO listed:
investigate coalescing the lifetimes into a single, shorter lifetime.